@@ -22,10 +22,10 @@ use Ogsteam\Ogspy\Model\Sessions_Model;
 
 
 /**
- * Stating an user Session
+ * Starting a user Session
  * @param $user_ip
  */
-function session_begin($user_ip)
+function session_begin($user_ip): void
 {
     global $cookie_id, $server_config, $pub_toolbar_type;
     $Sessions_Model = new Sessions_Model();
@@ -48,9 +48,9 @@ function session_begin($user_ip)
 /**
  * Gets the current session and creates it if the session for the current user does not exists
  */
-function session()
+function session(): void
 {
-    global $user_ip, $cookie_id, $server_config;
+    global $user_ip, $cookie_id, $server_config, $log;
     global $_COOKIE;
     $Sessions_Model = new Sessions_Model();
 
@@ -60,32 +60,46 @@ function session()
 
     //Purge des sessions expirées
     if ($server_config["session_time"] != 0) {
+        $log->debug("Cleaning expired sessions");
         $Sessions_Model->clean_expired_sessions();
     }
 
     //Récupération de l'id de session si cookie présent
     if (isset($_COOKIE[$cookie_name])) {
         $cookie_id = $_COOKIE[$cookie_name];
+        $log->debug("Session cookie found: " . substr($cookie_id, 0, 8) . "...");
 
         //Vérification de la validité de le session
         if (!$Sessions_Model->is_valid_session_id($cookie_id, $user_ip)) {
+            $log->warning("Invalid session for cookie_id: " . $cookie_id . " and user_ip: " . $user_ip);
+
             if (isset($server_config["disable_ip_check"]) && $server_config["disable_ip_check"] == 1) {
                 //Mise à jour de l'adresse ip de session si le contrôle des ip est désactivé
+                $log->info("Attempting to update IP address - IP check disabled");
+
                 if (!$Sessions_Model->update_session_public_ip($cookie_id, $user_ip)) {
+                    $log->warning("Failed to update session IP address");
                     $cookie_id = "";
+                } else {
+                    $log->info("Session IP address updated successfully");
                 }
             } else {
                 $cookie_id = "";
             }
+        } else {
+            $log->debug("Valid session found");
         }
+    } else {
+        $log->debug("No session cookie found");
     }
     if ($cookie_id == "") {
+        $log->info("Creating new session for IP address: " . $user_ip);
         session_begin($user_ip);
     } else {
         $cookie_expire = time() + $cookie_time * 60;
+        $log->debug("Updating session expiration time");
         $Sessions_Model->update_session_expiration_time($cookie_id, $cookie_expire);
     }
-
     session_set_user_data($cookie_id);
 }
 
@@ -94,64 +108,57 @@ function session()
  * @param int $user_id The current user
  * @param int $lastvisit Lastvisit timestamp
  */
-function session_set_user_id($user_id, $lastvisit = 0)
+function session_set_user_id($user_id, $lastvisit = 0): void
 {
-    global $user_ip, $cookie_id, $server_config;
+    global $user_ip, $cookie_id, $server_config,$log;
     $Sessions_Model = new Sessions_Model();
 
+    $log->info("Updating session for user ID: " . $user_id);
+
     if (isset($server_config["disable_ip_check"]) && $server_config["disable_ip_check"] != 1) {
+        $log->debug("Updating session with IP verification enabled");
         $Sessions_Model->update_session($user_id, $lastvisit, $cookie_id, $user_ip);
     } else {
+        $log->debug("Updating session without IP verification");
         $Sessions_Model->update_session($user_id, $lastvisit, $cookie_id);
     }
-
+    $log->debug("Retrieving user data after session update");
     session_set_user_data($cookie_id);
+    $log->info("User session ID: " . $user_id . " updated successfully");
 }
 
 /**
  * Set the user_data array according to the user parameters in the database
- * @param int $cookie_id The cookie id of the user
- * @todo Y a comme un probleme dans cette fonction... ne semble pas prendre de parametres alors que la fonction precedente lui en donne un...
+ * @param string $cookie_id The cookie id of the user
+ *
  */
-function session_set_user_data($cookie_id)
+function session_set_user_data(string $cookie_id)
 {
-    global $user_ip, $user_data, $user_auth, $user_token;
+    global $user_ip, $user_data, $user_auth, $user_token,$log;
 
+    $log->debug("Retrieving user data for cookie_id: " . substr($cookie_id, 0, 8) . "...");
     $user_data = (new Sessions_Model())->select_user_data_session($cookie_id, $user_ip);
 
-    if ($user_data == false) {
+    if (!$user_data) {
+        $log->warning("No user data found for cookie_id: " . substr($cookie_id, 0, 8) . "...");
         unset($user_data);
         unset($user_auth);
         unset($user_token);
+    } else {
+        if (isset($user_data["id"])) {
+            $log->info("User identified - ID: " . $user_data["id"] . ", username: " . $user_data["name"]);
+        } else {
+            $log->info("Unidentified visitor");
+        }
+
     }
 }
 
 /**
- * Get the user token (list) for the current user
- * @return mixed
- *
- *function session_set_user_tokens_data()
- *{
- *  global $db, $user_data, $user_token;
- *
- *  $request_tokens = "SELECT `name`,`token`,`expiration_date` FROM " . TABLE_USER_TOKEN . " WHERE `user_id` = " . $user_data["user_id"];
- *  $result_tokens = $db->sql_query($request_tokens);
- *
- *  if ($db->sql_numrows($result_tokens) > 0) {
- *
- *      $user_token = $db->sql_fetch_assoc($result_tokens);
- *  }
- *  else
- *  {
- *      $user_auth = user_get_auth($user_data["user_id"]);
- *  }
- *}*/
-
-/**
- * Closing an user session
+ * Closing a user session
  * @param boolean $user_id ID user session
  */
-function session_close($user_id = false)
+function session_close(bool $user_id = false): void
 {
     global $user_ip, $cookie_id, $server_config;
 
@@ -184,23 +191,21 @@ function session_whois_online()
 
     $tOnline = (new Sessions_Model())->who_is_online();
 
-    $guests = array();
-    $members = array();
+    $guests = [];
+    $members = [];
     foreach ($tOnline as $online) {
         $time_lastactivity = $online["session_expire"] - $cookie_time * 60;
         $session_ip = decode_ip($online["session_ip"]);
 
-        if (is_null($online["user_name"])) {
+        if (is_null($online["name"])) {
             $username = "Visiteur non identifié";
-            $guests[] = array("user" => $username, "time_start" => $online["session_start"], "time_lastactivity" => $time_lastactivity, "ip" => $session_ip, "ogs" => 0);
+            $guests[] = array("user" => $username, "time_start" => $online["session_start"], "time_lastactivity" => $time_lastactivity, "ip" => $session_ip, "session_type" => 0);
         } else {
-            $username = $online["user_name"];
-            $members[] = array("user" => $username, "time_start" => $online["session_start"], "time_lastactivity" => $time_lastactivity, "ip" => $session_ip, "ogs" => $online["session_ogs"]);
+            $username = $online["name"];
+            $members[] = array("user" => $username, "time_start" => $online["session_start"], "time_lastactivity" => $time_lastactivity, "ip" => $session_ip, "session_type" => $online["session_type"]);
         }
     }
-    $online = array_merge($members, $guests);
-
-    return $online;
+    return array_merge($members, $guests);
 }
 
 /**
