@@ -1,15 +1,6 @@
 <?php
 
-/**
- * Database Model
- *
- * @package OGSpy
- * @subpackage Model
- * @author DarkNoon
- * @copyright Copyright &copy; 2016, https://ogsteam.eu/
- * @license https://opensource.org/licenses/gpl-license.php GNU Public License
- * @version 3.4.0
- */
+
 
 namespace Ogsteam\Ogspy\Model;
 
@@ -18,72 +9,52 @@ use Ogsteam\Ogspy\Abstracts\Model_Abstract;
 class Spy_Model extends Model_Abstract
 {
     /**
-     * @param $user_id
-     * @param int $sort
-     * @param int $sort2
-     * @return array
+     * Retrieves a list of favorite spy reports for a specific user.
+     *
+     * @param int $user_id The ID of the user whose favorite spy reports are to be retrieved.
+     * @param int $sort Optional parameter to specify the sorting criterion.
+     *                  Values:
+     *                  1 = coordinates
+     *                  2 = ally
+     *                  3 = player
+     *                  4 = moon
+     *                  Default: 5 (dateRE).
+     * @param int $sort2 Optional parameter to specify the sorting order.
+     *                   Values:
+     *                   0 = descending (default).
+     *                   1 = ascending.
+     * @return array An associative array of favorite spy reports, where each report contains details such as
+     *               spy ID, galaxy, system, row, player, ally, moon, status, date, and poster name.
      */
-    public function get_favoriteSpyList($user_id, $sort = 5, $sort2 = 0)
+    public function get_favoriteSpyList(int $user_id, int $sort = 5, int $sort2 = 0)
     {
-        $user_id = (int)$user_id;
-        $sort = (int)$sort;
-        $sort2 = (int)$sort2;
+        $order = $sort2 === 0 ? " desc" : " asc";
+
+        $ordered_by = match ($sort) {
+                1 => "coordinates",
+                2 => "ally",
+                3 => "player",
+                4 => "moon",
+                default => "dateRE"
+            } . $order;
 
 
-        switch ($sort2) {
-            case 0:
-                $order = " desc";
-                break;
-            case 1:
-                $order = " asc";
-                break;
-            default:
-                $order = " asc";
-        }
+        $favorite = [];
 
-        switch ($sort) {
-            case 1:
-                $ordered_by = "coordinates" . $order . "";
-                break;
-            case 2:
-                $ordered_by = "ally " . $order;
-                break;
-            case 3:
-                $ordered_by = "player " . $order;
-                break;
-            case 4:
-                $ordered_by = "moon " . $order;
-                break;
-            case 5:
-                $ordered_by = "dateRE " . $order;
-                break;
-            default:
-                $ordered_by = "dateRE " . $order;
-        }
-
-
-        $favorite = array();
-
-        $request = "SELECT `id_spy`, `coordinates`, `dateRE`, `user`.`user_name`, `universe`.`moon`, `universe`.`ally`, `universe`.`player`, `universe`.`status`";
+        $request = "SELECT pspy.`id`, astro.`galaxy`, astro.`system`, astro.`row`, `dateRE`, `user`.`name`, `astro`.`type`, `player`.`ally_id`, `astro`.`player_id`, `player`.`status`";
         $request .= " FROM " . TABLE_PARSEDSPY. " `pspy`";
-        $request .= " INNER JOIN " . TABLE_UNIVERSE . " `universe` ON `pspy`.`coordinates` = CONCAT(`universe`.`galaxy`,':',`universe`.`system`,':',`universe`.`row`)";
-        $request .= " INNER JOIN " . TABLE_USER . " `user` ON `user`.`user_id` = `pspy`.`sender_id`";
+        $request .= " INNER JOIN " . TABLE_USER_BUILDING . " `astro` ON `pspy`.`astro_object_id` = `astro`.`id`";
+        $request .= " INNER JOIN " . TABLE_GAME_PLAYER . " `player` ON `astro`.`player_id` = `player`.`id`";
+        $request .= " INNER JOIN " . TABLE_USER . " `user` ON `user`.`id` = `pspy`.`sender_id`";
         $request .= " WHERE `pspy`.`sender_id`=$user_id ";
         $request .= " ORDER BY " . $ordered_by;
         $result = $this->db->sql_query($request);
 
-        while (list($spy_id, $coordinates, $datadate, $sender_name, $moon, $ally, $player, $status) = $this->db->sql_fetch_row($result)) {
+        while (list($spy_id, $galaxy, $system, $row, $datadate, $sender_name, $moon, $ally, $player, $status) = $this->db->sql_fetch_row($result)) {
 
             $favorite[$spy_id] = array(
-                "spy_id" => $spy_id, "spy_galaxy" => substr(
-                    $coordinates,
-                    0,
-                    strpos($coordinates, ':')
-                ), "spy_system" => substr($coordinates, strpos(
-                    $coordinates,
-                    ':'
-                ) + 1, strrpos($coordinates, ':') - strpos($coordinates, ':') - 1), "spy_row" =>
-                substr($coordinates, strrpos($coordinates, ':') + 1), "player" => $player,
+                "spy_id" => $spy_id, "spy_galaxy" => $galaxy
+                , "spy_system" =>$system, "spy_row" => $row, "player" => $player,
                 "ally" => $ally, "moon" => $moon, "status" => $status, "datadate" => $datadate,
                 "poster" => $sender_name
             );
@@ -92,28 +63,39 @@ class Spy_Model extends Model_Abstract
     }
 
 
-    public function get_spy_Id($id_RE)
+    /**
+     * Retrieves detailed information about a specific spy report.
+     *
+     * @param int $id_RE The unique identifier of the spy report to retrieve.
+     * @return array An associative array containing detailed information about the spy report,
+     *               including resources, activities, structures, fleet, technologies, probabilities,
+     *               and report date.
+     */
+    public function get_spy_Id(int $id_RE)
     {
-        $id_RE = (int)$id_RE;
-
-        $query = "SELECT `planet_name`, `coordinates`, `metal`, `cristal`, `deuterium`, `energie`, `activite`, `M`, `C`, `D`, `CES`, `CEF`, `UdR`, `UdN`, `CSp`, `HM`, `HC`,
+        $query = "SELECT `astro_object_id`, `metal`, `crystal`, `deuterium`, `energie`, `activite`, `M`, `C`, `D`, `CES`, `CEF`, `UdR`, `UdN`, `CSp`, `HM`, `HC`,
         `HD`, `Lab`, `Ter`, `Silo`, `Dock`, `DdR`, `BaLu`, `Pha`, `PoSa`, `LM`, `LLE`, `LLO`, `CG`, `AI`, `LP`, `PB`, `GB`, `MIC`, `MIP`, `PT`, `GT`, `CLE`, `CLO`, `CR`, `VB`, `VC`, `REC`, `SE`, `BMD`,
         `DST`, `EDLM`, `SAT`, `TRA`, `FOR`, `FAU`, `ECL`, `Esp`, `Ordi`, `Armes`, `Bouclier`, `Protection`, `NRJ`, `Hyp`, `RC`, `RI`, `PH`, `Laser`, `Ions`, `Plasma`, `RRI`, `Graviton`, `Astrophysique`,
-        `dateRE`, `proba` FROM " . TABLE_PARSEDSPY . " WHERE `id_spy`=$id_RE";
+        `dateRE`, `proba` FROM " . TABLE_PARSEDSPY . " WHERE `id`= {$id_RE}";
         $result = $this->db->sql_query($query);
 
-        $row = $this->db->sql_fetch_assoc($result);
-        return $row;
+        return $this->db->sql_fetch_assoc($result);
     }
 
-    public function get_all_spy_coordinates($coord)
+    /**
+     * Retrieves spy data for a specific astronomical object.
+     *
+     * @param int $astro_object_id The ID of the astronomical object for which the spy data is to be retrieved.
+     * @return array An array of associative arrays, where each entry contains detailed spy report information
+     *               such as planet name, resources (metal, crystal, deuterium, energy), activity, building levels,
+     *               defenses, fleet composition, technology levels, report date, and probability.
+     */
+    public function get_spy_data(int $astro_object_id)
     {
-        $coord = $this->db->sql_escape_string($coord);
-
-        $query = "SELECT `planet_name`, `coordinates`, `metal`, `cristal`, `deuterium`, `energie`, `activite`, `M`, `C`, `D`, `CES`, `CEF`, `UdR`, `UdN`, `CSp`, `HM`, `HC`,
+        $query = "SELECT `planet_name`, `astro_object_id`, `metal`, `crystal`, `deuterium`, `energie`, `activite`, `M`, `C`, `D`, `CES`, `CEF`, `UdR`, `UdN`, `CSp`, `HM`, `HC`,
         `HD`, `Lab`, `Ter`, `Silo`, `Dock`, `DdR`, `BaLu`, `Pha`, `PoSa`, `LM`, `LLE`, `LLO`, `CG`, `AI`, `LP`, `PB`, `GB`, `MIC`, `MIP`, `PT`, `GT`, `CLE`, `CLO`, `CR`, `VB`, `VC`, `REC`, `SE`, `BMD`,
         `DST`, `EDLM`, `SAT`, `TRA`, `FOR`, `FAU`, `ECL`, `Esp`, `Ordi`, `Armes`, `Bouclier`, `Protection`, `NRJ`, `Hyp`, `RC`, `RI`, `PH`, `Laser`, `Ions`, `Plasma`, `RRI`, `Graviton`, `Astrophysique`,
-        `dateRE`, `proba` FROM " . TABLE_PARSEDSPY . " WHERE `coordinates`='$coord' ORDER BY `dateRE` DESC ";
+        `dateRE`, `proba` FROM " . TABLE_PARSEDSPY . " WHERE `astro_object_id`='$astro_object_id' ORDER BY `dateRE` DESC ";
         $result = $this->db->sql_query($query);
 
         $tResult = array();
@@ -124,55 +106,60 @@ class Spy_Model extends Model_Abstract
     }
 
     /**
-     * @param int $galaxy
-     * @param int $system
-     * @param int $row
-     * @return int $nb_spy
+     * Retrieves the number of active spy reports associated with a specific planet based on its galaxy, system, and row coordinates.
+     *
+     * @param int $galaxy The galaxy number of the planet.
+     * @param int $system The system number of the planet.
+     * @param int $row The row number of the planet.
+     * @return int The count of active spy reports for the specified planet.
      */
-    public function get_nb_spy_by_planet($galaxy, $system, $row)
+    public function get_nb_spy_by_planet(int $galaxy, int $system, int $row)
     {
-        $galaxy = (int)$galaxy;
-        $system = (int)$system;
-        $row = (int)$row;
+        $request = "SELECT COUNT(pspy.id) as spy_count
+                FROM " . TABLE_PARSEDSPY . " pspy
+                INNER JOIN " . TABLE_USER_BUILDING . " astro ON pspy.astro_object_id = astro.id
+                AND astro.galaxy = " . $galaxy . "
+                AND astro.system = " . $system . "
+                AND astro.row = " . $row;
 
-        $request = "SELECT * FROM " . TABLE_PARSEDSPY . " WHERE `active` = '1' AND `coordinates` = '" . $galaxy . ":" . $system . ":" . $row . "'";
         $result = $this->db->sql_query($request);
-        $nb_spy = $this->db->sql_numrows($result);
+        $data = $this->db->sql_fetch_assoc($result);
+        $nb_spy = $data ? (int)$data['spy_count'] : 0;
 
         return $nb_spy;
     }
 
     /**
-     * @param int $galaxy
-     * @param int $system
-     * @param int $row
-     * @return array $tResult
+     * Retrieves a list of active spy reports for a specific astro object.
+     *
+     * @param int $astroObjectId The ID of the astro object for which active spy reports are to be retrieved.
+     * @return array An array of spy reports, where each report contains the spy ID, user name of the sender,
+     *               and the date of the report.
      */
-    public function get_spy_id_list_by_planet($galaxy, $system, $row)
+    public function get_spy_id_list_by_planet(int $astroObjectId)
     {
-        $galaxy = (int)$galaxy;
-        $system = (int)$system;
-        $row = (int)$row;
-
-        $request = "SELECT `id_spy`, `user_name`, `dateRE` "; //, `is_moon`";
-        $request .= " FROM " . TABLE_PARSEDSPY . " LEFT JOIN " . TABLE_USER . " ON `user_id` = `sender_id`";
-        $request .= " WHERE `active` = '1'  AND `coordinates` = '" . $galaxy . ":" . $system . ":" . $row . "'";
-        $request .= " ORDER BY `dateRE` DESC";
+        $request = "SELECT spy.`id`, u.`name`, spy.`dateRE` "; //, `is_moon`";
+        $request .= " FROM " . TABLE_PARSEDSPY . " spy LEFT JOIN " . TABLE_USER . " u ON u.`id` = spy.`sender_id`";
+        $request .= " WHERE spy.`active` = '1'  AND spy.`astro_object_id` = '$astroObjectId'";
+        $request .= " ORDER BY spy.`dateRE` DESC";
         $result = $this->db->sql_query($request);
         $tResult = array();
         while ($row = $this->db->sql_fetch_assoc($result)) {
             $tResult[] = array(
-                'id_spy' => $row['id_spy'],
-                'user_name' => $row['user_name'],
+                'spy_id' => $row['id'],
+                'user_name' => $row['name'],
                 'dateRE' => $row['dateRE']
             );
             //   ,"is_moon" => $row['is_moon']);
         }
-        return $tResult;
+        return  $tResult;
     }
 
     /**
-     * @param $spy_id
+     * Deletes a specific spy report based on its ID.
+     *
+     * @param int $spy_id The ID of the spy report to be deleted.
+     * @return void
      */
     public function delete_spy($spy_id)
     {
@@ -182,13 +169,14 @@ class Spy_Model extends Model_Abstract
         $this->db->sql_query($request);
 
         //todo prevoir suppression favorite ...
-
-
     }
 
     /**
-     * @param $spy_id
-     * @param $user_id
+     * Deletes a spy report from the database based on the given spy ID and sender ID.
+     *
+     * @param int $spy_id The unique identifier of the spy report to be deleted.
+     * @param int $user_id The ID of the sender associated with the spy report.
+     * @return void This method does not return any value.
      */
     public function delete_spy_by_senderId($spy_id, $user_id)
     {
@@ -200,10 +188,11 @@ class Spy_Model extends Model_Abstract
     }
 
 
-
     /**
-     * This function deletes expired Spy records from the database
-     * @param $limit_time
+     * Deletes expired spy reports from the database based on the provided time limit or inactive status.
+     *
+     * @param int $limit_time The timestamp used to determine expiration. Spy reports with a date earlier than this value, or marked as inactive, will be deleted.
+     * @return void
      */
     public function delete_expired_spies($limit_time)
     {
