@@ -1,4 +1,4 @@
-<?php
+<?php global $lang;
 /**
  * Panneau administration des options Modules
  * @package OGSpy
@@ -19,11 +19,11 @@ $mod_list = mod_list();
 <table class="og-table og-medium-table">
     <thead>
         <tr>
-            <th  colspan="6" ><?php echo ($lang['ADMIN_MOD_LIST']); ?></th>
+            <th  colspan="6" ><?= $lang['ADMIN_MOD_LIST'] ?></th>
         </tr>
         <tr>
-            <th  colspan="5" ><?php echo ($lang['ADMIN_MOD_USER']); ?></th>
-            <th><?php echo ($lang['ADMIN_MOD_MENUVIEW']); ?></th>
+            <th  colspan="5" ><?= ($lang['ADMIN_MOD_USER']) ?></th>
+            <th><?= ($lang['ADMIN_MOD_MENUVIEW']) ?></th>
         </tr>
     </thead>
     <tbody>
@@ -32,39 +32,39 @@ $mod_list = mod_list();
             <?php if ($mod["admin_only"] == 0) : ?>
                 <tr>
                     <td>
-                        <?php echo "(" . $mod['position'] . ') ' . $mod["title"] . " (" . $mod["version"] . ")"; ?>
+                        <?= "(" . $mod['position'] . ') ' . $mod["title"] . " (" . $mod["version"] . ")" ?>
                     </td>
                     <td>
-                        <a href='index.php?action=mod_up&amp;mod_id=<?php echo $mod['id']; ?>'>
+                        <a href='index.php?action=mod_up&amp;mod_id=<?= $mod['id'] ?>'>
                             <img src='images/asc.png' title='Monter'>
                         </a>
                         &nbsp;
-                        <a href='index.php?action=mod_down&amp;mod_id=<?php echo $mod['id']; ?>'>
+                        <a href='index.php?action=mod_down&amp;mod_id=<?= $mod['id'] ?>'>
                             <img src='images/desc.png' title='Descendre'>
                         </a>
                     </td>
                     <td>
-                        <a href='index.php?action=mod_disable&amp;mod_id=<?php echo $mod['id']; ?>'>
-                            <?php echo $lang['ADMIN_MOD_DISABLE']; ?>
+                        <a href='index.php?action=mod_disable&amp;mod_id=<?= $mod['id'] ?>'>
+                            <?= $lang['ADMIN_MOD_DISABLE'] ?>
                         </a>
                     </td>
                     <td>
-                        <a href='index.php?action=mod_uninstall&amp;mod_id=<?php echo $mod['id']; ?>'>
-                            <?php echo $lang['ADMIN_MOD_REMOVE']; ?>
+                        <a href='index.php?action=mod_uninstall&amp;mod_id=<?= $mod['id'] ?>'>
+                            <?= $lang['ADMIN_MOD_REMOVE'] ?>
                         </a>
                     </td>
 
                     <td>
                         <?php if (!$mod["up_to_date"]) : ?>
-                            <a href='index.php?action=mod_update&amp;mod_id=<?php echo $mod['id']; ?>'>
-                                <?php echo $lang['ADMIN_MOD_UPDATE']; ?>
+                            <a href='index.php?action=mod_update&amp;mod_id=<?= $mod['id'] ?>'>
+                                <?= $lang['ADMIN_MOD_UPDATE'] ?>
                             </a>
                         <?php endif; ?>
                     </td>
 
                     <td>
-                        <a href='index.php?action=mod_admin&amp;mod_id=<?php echo $mod['id']; ?>'>
-                            <?php echo $lang['ADMIN_MOD_NORMAL']; ?>
+                        <a href='index.php?action=mod_admin&amp;mod_id=<?= $mod['id'] ?>'>
+                            <?= $lang['ADMIN_MOD_NORMAL'] ?>
                         </a>
                     </td>
                 </tr>
@@ -74,8 +74,8 @@ $mod_list = mod_list();
     <thead>
 
         <tr>
-            <th  colspan="5" ><?php echo ($lang['ADMIN_MOD_ADMIN']); ?></th>
-            <th><?php echo ($lang['ADMIN_MOD_MENUVIEW']); ?></th>
+            <th  colspan="5" ><?= ($lang['ADMIN_MOD_ADMIN']) ?></th>
+            <th><?= ($lang['ADMIN_MOD_MENUVIEW']) ?></th>
         </tr>
     </thead>
     <tbody>
@@ -83,39 +83,39 @@ $mod_list = mod_list();
             <?php if ($mod["admin_only"] == 1) : ?>
                 <tr>
                     <td>
-                        <?php echo "(" . $mod['position'] . ') ' . $mod["title"] . " (" . $mod["version"] . ")"; ?>
+                        <?= "(" . $mod['position'] . ') ' . $mod["title"] . " (" . $mod["version"] . ")" ?>
                     </td>
                     <td>
-                        <a href='index.php?action=mod_up&amp;mod_id=<?php echo $mod['id']; ?>'>
+                        <a href='index.php?action=mod_up&amp;mod_id=<?= $mod['id'] ?>'>
                             <img src='images/asc.png' title='Monter'>
                         </a>
                         &nbsp;
-                        <a href='index.php?action=mod_down&amp;mod_id=<?php echo $mod['id']; ?>'>
+                        <a href='index.php?action=mod_down&amp;mod_id=<?= $mod['id'] ?>'>
                             <img src='images/desc.png' title='Descendre'>
                         </a>
                     </td>
                     <td>
-                        <a href='index.php?action=mod_disable&amp;mod_id=<?php echo $mod['id']; ?>'>
-                            <?php echo $lang['ADMIN_MOD_DISABLE']; ?>
+                        <a href='index.php?action=mod_disable&amp;mod_id=<?= $mod['id'] ?>'>
+                            <?= $lang['ADMIN_MOD_DISABLE'] ?>
                         </a>
                     </td>
                     <td>
-                        <a href='index.php?action=mod_uninstall&amp;mod_id=<?php echo $mod['id']; ?>'>
-                            <?php echo $lang['ADMIN_MOD_REMOVE']; ?>
+                        <a href='index.php?action=mod_uninstall&amp;mod_id=<?= $mod['id'] ?>'>
+                            <?= $lang['ADMIN_MOD_REMOVE'] ?>
                         </a>
                     </td>
 
                     <td>
                         <?php if (!$mod["up_to_date"]) : ?>
-                            <a href='index.php?action=mod_update&amp;mod_id=<?php echo $mod['id']; ?>'>
-                                <?php echo $lang['ADMIN_MOD_UPDATE']; ?>
+                            <a href='index.php?action=mod_update&amp;mod_id=<?= $mod['id'] ?>'>
+                                <?= $lang['ADMIN_MOD_UPDATE'] ?>
                             </a>
                         <?php endif; ?>
                     </td>
 
                     <td>
-                        <a href='index.php?action=mod_admin&amp;mod_id=<?php echo $mod['id']; ?>'>
-                            <?php echo $lang['ADMIN_MOD_ADMIN']; ?>
+                        <a href='index.php?action=mod_admin&amp;mod_id=<?= $mod['id'] ?>'>
+                            <?= $lang['ADMIN_MOD_ADMIN'] ?>
                         </a>
                     </td>
                 </tr>
@@ -130,7 +130,7 @@ $mod_list = mod_list();
     <table class="og-table og-medium-table">
         <thead>
             <tr>
-                <th  colspan="4" ><?php echo ($lang['ADMIN_MOD_LIST_INACTIVE']); ?></th>
+                <th  colspan="4" ><?= ($lang['ADMIN_MOD_LIST_INACTIVE']) ?></th>
             </tr>
         </thead>
         <tbody>
@@ -138,22 +138,22 @@ $mod_list = mod_list();
             <?php foreach ($mods as $mod) : ?>
                 <tr>
                     <td>
-                        <?php echo $mod["title"] . " (" . $mod["version"] . ")"; ?>
+                        <?= $mod["title"] . " (" . $mod["version"] . ")" ?>
                     </td>
                     <td>
-                        <a href='index.php?action=mod_active&amp;mod_id=<?php echo $mod['id']; ?>'>
-                            <?php echo $lang['ADMIN_MOD_ENABLE']; ?>
+                        <a href='index.php?action=mod_active&amp;mod_id=<?= $mod['id'] ?>'>
+                            <?= $lang['ADMIN_MOD_ENABLE'] ?>
                         </a>
                     </td>
                     <td>
-                        <a href='index.php?action=mod_uninstall&amp;mod_id=<?php echo $mod['id']; ?>'>
-                            <?php echo $lang['ADMIN_MOD_REMOVE']; ?>
+                        <a href='index.php?action=mod_uninstall&amp;mod_id=<?= $mod['id'] ?>'>
+                            <?= $lang['ADMIN_MOD_REMOVE'] ?>
                         </a>
                     </td>
                     <td>
                         <?php if (!$mod["up_to_date"]) : ?>
-                            <a href='index.php?action=mod_update&amp;mod_id=<?php echo $mod['id']; ?>'>
-                                <?php echo $lang['ADMIN_MOD_UPDATE']; ?>
+                            <a href='index.php?action=mod_update&amp;mod_id=<?= $mod['id'] ?>'>
+                                <?= $lang['ADMIN_MOD_UPDATE'] ?>
                             </a>
                         <?php endif; ?>
                     </td>
@@ -169,18 +169,18 @@ $mod_list = mod_list();
     <table class="og-table og-medium-table">
         <thead>
             <tr>
-                <th  colspan="2" ><?php echo ($lang['ADMIN_MOD_NOT_INSTALLED']); ?></th>
+                <th  colspan="2" ><?= ($lang['ADMIN_MOD_NOT_INSTALLED']) ?></th>
             </tr>
         </thead>
         <tbody>
             <?php foreach ($mods as $mod) : ?>
                 <tr>
                     <td>
-                        <?php echo $mod["title"]; ?>
+                        <?= $mod["title"] ?>
                     </td>
                     <td>
-                        <a href='index.php?action=mod_install&amp;directory=<?php echo $mod['directory']; ?>'>
-                            <?php echo $lang['ADMIN_MOD_INSTALL']; ?>
+                        <a href='index.php?action=mod_install&amp;directory=<?= $mod['directory'] ?>'>
+                            <?= $lang['ADMIN_MOD_INSTALL'] ?>
                         </a>
                     </td>
                 </tr>
@@ -196,18 +196,18 @@ $mod_list = mod_list();
     <table class="og-table og-medium-table">
         <thead>
             <tr>
-                <th  colspan="2" ><?php echo ($lang['ADMIN_MOD_INVALID']); ?></th>
+                <th  colspan="2" ><?= ($lang['ADMIN_MOD_INVALID']) ?></th>
             </tr>
         </thead>
         <tbody>
             <?php foreach ($mods as $mod) : ?>
                 <tr>
                     <td>
-                        <?php echo $mod["title"]; ?>
+                        <?= $mod["title"] ?>
                     </td>
                     <td>
-                        <a href='index.php?action=mod_uninstall&amp;mod_id=<?php echo $mod['id']; ?>'>
-                            <?php echo $lang['ADMIN_MOD_REMOVE']; ?>
+                        <a href='index.php?action=mod_uninstall&amp;mod_id=<?= $mod['id'] ?>'>
+                            <?= $lang['ADMIN_MOD_REMOVE'] ?>
                         </a>
                     </td>
                 </tr>
