@@ -1,4 +1,4 @@
-<?php
+<?php global $user_data, $server_config, $lang;
 /**
  * Panneau d'Administration : paramètres et options du serveur
  * @package OGSpy
@@ -13,13 +13,13 @@ if (!defined('IN_SPYOGAME')) {
     die("Hacking attempt");
 }
 
-if ($user_data["user_admin"] != 1 && $user_data["user_coadmin"] != 1) {
+if ($user_data["admin"] != 1 && $user_data["coadmin"] != 1) {
     redirection("index.php?action=message&amp;id_message=forbidden&amp;info");
 }
 
 //check mail
 if (isset($pub_testmail)) {
-    sendMail($user_data["user_email"], "TEST", "<h1>TEST OK</h1>");
+    sendMail($user_data["email"], "TEST", "<h1>TEST OK</h1>");
 }
 
 
@@ -45,7 +45,7 @@ $num_of_galaxies = (isset($pub_num_of_galaxies)) ? $pub_num_of_galaxies : $serve
 $num_of_systems = (isset($pub_num_of_systems)) ? $pub_num_of_systems : $server_config['num_of_systems'];
 $block_ratio = $server_config['block_ratio'] == 1 ? "checked" : "";
 $ratio_limit = $server_config['ratio_limit'];
-$speed_uni = $server_config['speed_uni'];
+$speed_uni = $server_config['speed_uni'] ?? 1;
 $speed_fleet_peaceful = $server_config['speed_fleet_peaceful'];
 $speed_fleet_war = $server_config['speed_fleet_war'];
 $speed_fleet_holding = $server_config['speed_fleet_holding'];
@@ -78,271 +78,271 @@ $mail_smtp_password = "";
         <thead>
             <tr>
                 <th colspan="2">
-                    <?php echo ($lang['ADMIN_PARAMS_GENERAL']); ?>
+                    <?= $lang['ADMIN_PARAMS_GENERAL'] ?>
                 </th>
             </tr>
         </thead>
         <tbody>
             <tr>
-                <td class="tdstat"><?php echo ($lang['ADMIN_PARAMS_SERVERNAME']); ?></td>
-                <td class="tdvalue"><input type="text" name="servername" size="60"  value="<?php echo $servername; ?>"></td>
+                <td class="tdstat"><?= ($lang['ADMIN_PARAMS_SERVERNAME']) ?></td>
+                <td class="tdvalue"><input type="text" name="servername" size="60"  value="<?= $servername ?>"></td>
             </tr>
             <tr>
-                <td class="tdstat"><?php echo ($lang['ADMIN_PARAMS_ACTIVATESERVER']); ?><?php echo help("admin_server_status"); ?></td>
-                <td class="tdvalue"><input name="server_active" type="checkbox" value="1" <?php echo $server_active; ?>></td>
+                <td class="tdstat"><?= ($lang['ADMIN_PARAMS_ACTIVATESERVER']) ?><?= help("admin_server_status") ?></td>
+                <td class="tdvalue"><input name="server_active" type="checkbox" value="1" <?= $server_active ?>></td>
             </tr>
             <tr>
-                <td class="tdstat"> <?php echo ($lang['ADMIN_PARAMS_OFFREASON']); ?><?php echo help("admin_server_status_message"); ?></td>
-                <td class="tdvalue"><input type="text" name="reason" size="60" value="<?php echo $reason; ?>"></td>
+                <td class="tdstat"> <?= ($lang['ADMIN_PARAMS_OFFREASON']) ?><?= help("admin_server_status_message") ?></td>
+                <td class="tdvalue"><input type="text" name="reason" size="60" value="<?= $reason ?>"></td>
             </tr>
         </tbody>
         <thead>
             <tr>
                 <th colspan="2">
-                    <?php echo ($lang['ADMIN_PARAMS_MEMBEROPTIONS']); ?>
+                    <?= ($lang['ADMIN_PARAMS_MEMBEROPTIONS']) ?>
                 </th>
             </tr>
         </thead>
         <tbody>
             <tr>
-                <td class="tdstat"><?php echo ($lang['ADMIN_PARAMS_ALLOW_IPCHECKDISABLING']); ?><?php echo help("admin_check_ip"); ?></td>
-                <td class="tdvalue"><input name="disable_ip_check" type="checkbox" value="1" <?php echo $disable_ip_check; ?>></td>
+                <td class="tdstat"><?= ($lang['ADMIN_PARAMS_ALLOW_IPCHECKDISABLING']) ?><?= help("admin_check_ip") ?></td>
+                <td class="tdvalue"><input name="disable_ip_check" type="checkbox" value="1" <?= $disable_ip_check ?>></td>
             </tr>
             <tr>
-                <td class="tdstat"><?php echo ($lang['ADMIN_PARAMS_MAXSSFAVORITES']); ?></td>
-                <td class="tdvalue"><input name="max_favorites" type="text" size="5" maxlength="2" value="<?php echo $max_favorites; ?>">
+                <td class="tdstat"><?= ($lang['ADMIN_PARAMS_MAXSSFAVORITES']) ?></td>
+                <td class="tdvalue"><input name="max_favorites" type="text" size="5" maxlength="2" value="<?= $max_favorites ?>">
                 </td>
             </tr>
             <tr>
-                <td class="tdstat"><?php echo ($lang['ADMIN_PARAMS_MAXREFAVORITES']); ?></td>
-                <td class="tdvalue"><input name="max_favorites_spy" type="text" size="5" maxlength="2" value="<?php echo $max_favorites_spy; ?>"></td>
+                <td class="tdstat"><?= ($lang['ADMIN_PARAMS_MAXREFAVORITES']) ?></td>
+                <td class="tdvalue"><input name="max_favorites_spy" type="text" size="5" maxlength="2" value="<?= $max_favorites_spy ?>"></td>
             </tr>
         </tbody>
         <thead>
             <tr>
                 <th colspan="2">
-                    <?php echo ($lang['ADMIN_PARAMS_SESSIONS_TITLE']); ?>
+                    <?= ($lang['ADMIN_PARAMS_SESSIONS_TITLE']) ?>
                 </th>
             </tr>
         </thead>
         <tbody>
             <tr>
-                <td class="tdstat"><?php echo ($lang['ADMIN_PARAMS_SESSIONS_DURATION']); ?><?php echo help("admin_session_infini"); ?></th>
-                <td class="tdvalue"><input name="session_time" type="text" size="5" maxlength="3" value="<?php echo $session_time; ?>"></th>
+                <td class="tdstat"><?= ($lang['ADMIN_PARAMS_SESSIONS_DURATION']) ?><?= help("admin_session_infini") ?></td>
+                <td class="tdvalue"><input name="session_time" type="text" size="5" maxlength="3" value="<?= $session_time ?>"></td>
             </tr>
         </tbody>
         <thead>
             <tr>
                 <th colspan="2">
-                    <?php echo ($lang['ADMIN_PARAMS_ALLYPROTECT']); ?>
+                    <?= ($lang['ADMIN_PARAMS_ALLYPROTECT']) ?>
                 </th>
             </tr>
         </thead>
         <tbody>
             <tr>
-                <td class="tdstat"><?php echo ($lang['ADMIN_PARAMS_ALLYPROTECTLIST']); ?><br />
+                <td class="tdstat"><?= ($lang['ADMIN_PARAMS_ALLYPROTECTLIST']) ?><br />
 
-                    <i class="og-warning"><?php echo ($lang['ADMIN_PARAMS_ALLYPROTECTNOTICE']); ?></i>
+                    <i class="og-warning"><?= ($lang['ADMIN_PARAMS_ALLYPROTECTNOTICE']) ?></i>
                 </td>
-                <td class="tdvalue"><input type="text" size="60" name="ally_protection" value="<?php echo $ally_protection; ?>"></td>
+                <td class="tdvalue"><input type="text" size="60" name="ally_protection" value="<?= $ally_protection ?>"></td>
             </tr>
             <tr>
-                <td class="tdstat"><?php echo ($lang['ADMIN_PARAMS_ALLYPROTECTFRIENDS']); ?>
+                <td class="tdstat"><?= ($lang['ADMIN_PARAMS_ALLYPROTECTFRIENDS']) ?>
                     <br />
-                    <i class="og-warning"><?php echo ($lang['ADMIN_PARAMS_ALLYPROTECTNOTICE']); ?></i>
+                    <i class="og-warning"><?= ($lang['ADMIN_PARAMS_ALLYPROTECTNOTICE']) ?></i>
                 </td>
-                <td class="tdvalue"><input type="text" size="60" name="allied" value="<?php echo $allied; ?>"></td>
+                <td class="tdvalue"><input type="text" size="60" name="allied" value="<?= $allied ?>"></td>
             </tr>
         </tbody>
         <thead>
             <tr>
-                <th colspan="2"><?php echo ($lang['ADMIN_PARAMS_OTHER']); ?></th>
-            </tr> 
+                <th colspan="2"><?= ($lang['ADMIN_PARAMS_OTHER']) ?></th>
+            </tr>
         </thead>
         <tbody>
             <tr>
-                <td class="tdstat"><?php echo ($lang['ADMIN_PARAMS_FORUMLINK']); ?></td>
-                <td class="tdvalue"><input type="text" size="60" name="url_forum" value="<?php echo $url_forum; ?>"></td>
+                <td class="tdstat"><?= ($lang['ADMIN_PARAMS_FORUMLINK']) ?></td>
+                <td class="tdvalue"><input type="text" size="60" name="url_forum" value="<?= $url_forum ?>"></td>
             </tr>
             <tr>
-                <td class="tdstat"><?php echo ($lang['ADMIN_PARAMS_DEBUGSQL']); ?><?php echo help("admin_save_transaction"); ?><br />
-                    <i class="og-warning"><?php echo ($lang['ADMIN_PARAMS_DEBUGSQLALERT']); ?></i>
+                <td class="tdstat"><?= ($lang['ADMIN_PARAMS_DEBUGSQL']) ?><?= help("admin_save_transaction") ?><br />
+                    <i class="og-warning"><?= ($lang['ADMIN_PARAMS_DEBUGSQLALERT']) ?></i>
                 </td>
-                <td class="tdvalue"><input name="debug_log" type="checkbox" value="1" <?php echo $debug_log; ?>></td>
+                <td class="tdvalue"><input name="debug_log" type="checkbox" value="1" <?= $debug_log ?>></td>
             </tr>
             <tr>
-                <td class="tdstat"><?php echo ($lang['ADMIN_PARAMS_RATIOMOD']); ?></td>
-                <td class="tdvalue"><input name="block_ratio" type="checkbox" value="1" <?php echo $block_ratio; ?>></td>
+                <td class="tdstat"><?= ($lang['ADMIN_PARAMS_RATIOMOD']) ?></td>
+                <td class="tdvalue"><input name="block_ratio" type="checkbox" value="1" <?= $block_ratio ?>></td>
             </tr>
             <tr>
-                <td class="tdstat"><?php echo ($lang['ADMIN_PARAMS_RATIOBLIMIT']); ?></td>
-                <td class="tdvalue"><input name="ratio_limit" type="number" size="10" maxlength="9" value="<?php echo $ratio_limit; ?>"></td>
+                <td class="tdstat"><?= ($lang['ADMIN_PARAMS_RATIOBLIMIT']) ?></td>
+                <td class="tdvalue"><input name="ratio_limit" type="number" size="10" maxlength="9" value="<?= $ratio_limit ?>"></td>
             </tr>
         </tbody>
         <thead>
             <tr>
-                <th colspan="2"><?php echo ($lang['ADMIN_PARAMS_MAIL']); ?></th>
+                <th colspan="2"><?= ($lang['ADMIN_PARAMS_MAIL']) ?></th>
             </tr>
         </thead>
         <tbody>
             <tr>
-                <td class="tdstat"><?php echo ($lang['ADMIN_PARAMS_MAIL_USE']); ?></td>
-                <td class="tdvalue"><input name="mail_use" type="checkbox" value="1" <?php echo $mail_use; ?>></td>
+                <td class="tdstat"><?= ($lang['ADMIN_PARAMS_MAIL_USE']) ?></td>
+                <td class="tdvalue"><input name="mail_use" type="checkbox" value="1" <?= $mail_use ?>></td>
             </tr>
             <tr>
-                <td class="tdstat"><?php echo ($lang['ADMIN_PARAMS_MAIL_SMTP_USE']); ?></td>
-                <td class="tdvalue"><input name="mail_smtp_use" type="checkbox" value="1" <?php echo $mail_smtp_use; ?>></td>
+                <td class="tdstat"><?= ($lang['ADMIN_PARAMS_MAIL_SMTP_USE']) ?></td>
+                <td class="tdvalue"><input name="mail_smtp_use" type="checkbox" value="1" <?= $mail_smtp_use ?>></td>
             </tr>
             <tr>
-                <td class="tdstat"><?php echo ($lang['ADMIN_PARAMS_MAIL_SMTP_SECURE']); ?></td>
-                <td class="tdvalue"><input name="mail_smtp_secure" type="checkbox" value="1" <?php echo $mail_smtp_secure; ?>></td>
+                <td class="tdstat"><?= ($lang['ADMIN_PARAMS_MAIL_SMTP_SECURE']) ?></td>
+                <td class="tdvalue"><input name="mail_smtp_secure" type="checkbox" value="1" <?= $mail_smtp_secure ?>></td>
             </tr>
             <tr>
-                <td class="tdstat"><?php echo ($lang['ADMIN_PARAMS_MAIL_SMTP_HOST']); ?></td>
-                <td class="tdvalue"><input type="text" size="60" name="mail_smtp_host" value="<?php echo $mail_smtp_host; ?>"></td>
+                <td class="tdstat"><?= ($lang['ADMIN_PARAMS_MAIL_SMTP_HOST']) ?></td>
+                <td class="tdvalue"><input type="text" size="60" name="mail_smtp_host" value="<?= $mail_smtp_host ?>"></td>
             </tr>
             <tr>
-                <td class="tdstat"><?php echo ($lang['ADMIN_PARAMS_MAIL_SMTP_PORT']); ?></td>
-                <td class="tdvalue"><input type="text" name="mail_smtp_port" maxlength="4" size="5" value="<?php echo $mail_smtp_port; ?>"></td>
+                <td class="tdstat"><?= ($lang['ADMIN_PARAMS_MAIL_SMTP_PORT']) ?></td>
+                <td class="tdvalue"><input type="text" name="mail_smtp_port" maxlength="4" size="5" value="<?= $mail_smtp_port ?>"></td>
             </tr>
             <tr>
-                <td class="tdstat"><?php echo ($lang['ADMIN_PARAMS_MAIL_SMTP_USERNAME']); ?></td>
-                <td class="tdvalue"><input type="text" size="30" name="mail_smtp_username" value="<?php echo $mail_smtp_username; ?>"></td>
+                <td class="tdstat"><?= ($lang['ADMIN_PARAMS_MAIL_SMTP_USERNAME']) ?></td>
+                <td class="tdvalue"><input type="text" size="30" name="mail_smtp_username" value="<?= $mail_smtp_username ?>"></td>
             </tr>
             <tr>
-                <td class="tdstat""><?php echo ($lang['ADMIN_PARAMS_MAIL_SMTP_PASSEWORD']); ?></td>
-                <td class="tdvalue"><input type="password" size="30" name="mail_smtp_password" value="<?php echo $mail_smtp_password; ?>"> (<input type="checkbox" name="enable_mail_smtp_password" />)</td>
+                <td class="tdstat""><?= ($lang['ADMIN_PARAMS_MAIL_SMTP_PASSEWORD']) ?></td>
+                <td class="tdvalue"><input type="password" size="30" name="mail_smtp_password" value="<?= $mail_smtp_password ?>"> (<input type="checkbox" name="enable_mail_smtp_password" />)</td>
             </tr>
-            <?php if ($server_config['mail_use'] == 1 && check_var($user_data["user_email"], "Email")) : ?>
+            <?php if ($server_config['mail_use'] == 1 && check_var($user_data["email"], "Email")) : ?>
                 <tr>
-                    <td class="tdstat"><?php echo ($lang['ADMIN_PARAMS_MAIL_TEST'] . $user_data["user_email"]); ?></td>
+                    <td class="tdstat"><?= ($lang['ADMIN_PARAMS_MAIL_TEST'] . $user_data["email"]) ?></td>
                     <td class="tdvalue"><a href="index.php?action=administration&subaction=parameter&testmail">TEST</a></td>
                 </tr>
             <?php endif; ?>
         </tbody>
         <thead>
             <tr>
-                <th colspan="2"><?php echo ($lang['ADMIN_PARAMS_SERVICE']); ?></th>
-            </tr>  
+                <th colspan="2"><?= ($lang['ADMIN_PARAMS_SERVICE']) ?></th>
+            </tr>
         </thead>
         <tbody>
             <tr>
-                <td class="tdstat"><?php echo ($lang['ADMIN_PARAMS_DURATION_RANKS']); ?></td>
-                <td class="tdvalue"><input type="text" name="max_keeprank" maxlength="4" size="5" value="<?php echo $max_keeprank; ?>">&nbsp;<select name="keeprank_criterion">
-                        <option value="quantity" <?php echo $keeprank_criterion == "quantity" ? "selected" : ""; ?>><?php echo ($lang['ADMIN_PARAMS_DURATION_NUMBER']); ?></option>
-                        <option value="day" <?php echo $keeprank_criterion == "day" ? "selected" : ""; ?>><?php echo ($lang['ADMIN_PARAMS_DURATION_DAYS']); ?></option>
+                <td class="tdstat"><?= ($lang['ADMIN_PARAMS_DURATION_RANKS']) ?></td>
+                <td class="tdvalue"><input type="text" name="max_keeprank" maxlength="4" size="5" value="<?= $max_keeprank ?>">&nbsp;<select name="keeprank_criterion">
+                        <option value="quantity" <?= $keeprank_criterion == "quantity" ? "selected" : "" ?>><?= ($lang['ADMIN_PARAMS_DURATION_NUMBER']) ?></option>
+                        <option value="day" <?= $keeprank_criterion == "day" ? "selected" : "" ?>><?= ($lang['ADMIN_PARAMS_DURATION_DAYS']) ?></option>
                     </select></td>
             </tr>
             <tr>
-                <td class="tdstat"><?php echo ($lang['ADMIN_PARAMS_MAX_SPYREPORTS']); ?></td>
-                <td class="tdvalue"><input type="text" name="max_spyreport" maxlength="4" size="5" value="<?php echo $max_spyreport; ?>">
+                <td class="tdstat"><?= ($lang['ADMIN_PARAMS_MAX_SPYREPORTS']) ?></td>
+                <td class="tdvalue"><input type="text" name="max_spyreport" maxlength="4" size="5" value="<?= $max_spyreport ?>">
                 </td>
             </tr>
             <tr>
-                <td class="tdstat"><?php echo ($lang['ADMIN_PARAMS_DURATION_SPYREPORTS']); ?></td>
-                <td class="tdvalue"><input type="text" name="max_keepspyreport" maxlength="4" size="5" value="<?php echo $max_keepspyreport; ?>"></td>
+                <td class="tdstat"><?= ($lang['ADMIN_PARAMS_DURATION_SPYREPORTS']) ?></td>
+                <td class="tdvalue"><input type="text" name="max_keepspyreport" maxlength="4" size="5" value="<?= $max_keepspyreport ?>"></td>
             </tr>
             <tr>
-                <td class="tdstat"><?php echo ($lang['ADMIN_PARAMS_DURATION_LOGS']); ?></td>
-                <td class="tdvalue"><input name="max_keeplog" type="text" size="5" maxlength="3" value="<?php echo $max_keeplog; ?>"></td>
-            </tr>     
+                <td class="tdstat"><?= ($lang['ADMIN_PARAMS_DURATION_LOGS']) ?></td>
+                <td class="tdvalue"><input name="max_keeplog" type="text" size="5" maxlength="3" value="<?= $max_keeplog ?>"></td>
+            </tr>
         </tbody>
-        <?php if ($user_data["user_admin"] == 1) : ?>
+        <?php if ($user_data["admin"] == 1) : ?>
             <thead>
                 <tr>
-                    <th colspan="2"><?php echo ($lang['ADMIN_PARAMS_GAME_OPTIONS']); ?></th>
+                    <th colspan="2"><?= ($lang['ADMIN_PARAMS_GAME_OPTIONS']) ?></th>
                 </tr>
             </thead>
             <tbody>
                 <tr>
-                    <td class="tdstat"><?php echo ($lang['ADMIN_PARAMS_GAME_GALAXIES']); ?><?php echo help("profile_galaxy"); ?></td>
-                    <td class="tdvalue"><input name="num_of_galaxies" id="galaxies" type="text" size="5" maxlength="3" value="<?php echo $num_of_galaxies; ?>" onChange="if (!confirm('<?php echo ($lang['ADMIN_PARAMS_GAME_GALAXIES_POPUP']); ?>')) {
-                                document.getElementById('galaxies').value = '<?php echo $num_of_galaxies; ?>';
+                    <td class="tdstat"><?= ($lang['ADMIN_PARAMS_GAME_GALAXIES']) ?><?= help("profile_galaxy") ?></td>
+                    <td class="tdvalue"><input name="num_of_galaxies" id="galaxies" type="text" size="5" maxlength="3" value="<?= $num_of_galaxies ?>" onChange="if (!confirm('<?= ($lang['ADMIN_PARAMS_GAME_GALAXIES_POPUP']) ?>')) {
+                                document.getElementById('galaxies').value = '<?= $num_of_galaxies ?>';
                             }" readonly="readonly">(<input name="enable_input_num_galaxies" type="checkbox" onClick="(this.checked) ? document.getElementById('galaxies').readOnly = false : document.getElementById('galaxies').readOnly = true;">)
                     </td>
                 </tr>
                 <tr>
-                    <td class="tdstat"><?php echo ($lang['ADMIN_PARAMS_GAME_SYSTEMS']); ?><?php echo help("profile_galaxy"); ?></td>
-                    <td class="tdvalue"><input name="num_of_systems" id="systems" type="text" size="5" maxlength="3" value="<?php echo $num_of_systems; ?>" onChange="if (!confirm('<?php echo ($lang['ADMIN_PARAMS_GAME_SYSTEMS_POPUP']); ?>')) {
-                                document.getElementById('systems').value = '<?php echo $num_of_systems; ?>';
+                    <td class="tdstat"><?= ($lang['ADMIN_PARAMS_GAME_SYSTEMS']) ?><?= help("profile_galaxy") ?></td>
+                    <td class="tdvalue"><input name="num_of_systems" id="systems" type="text" size="5" maxlength="3" value="<?= $num_of_systems ?>" onChange="if (!confirm('<?= ($lang['ADMIN_PARAMS_GAME_SYSTEMS_POPUP']) ?>')) {
+                                document.getElementById('systems').value = '<?= $num_of_systems ?>';
                             }" readonly="readonly">(<input name="enable_input_num_systems" type="checkbox" onClick="(this.checked) ? document.getElementById('systems').readOnly = false : document.getElementById('systems').readOnly = true;">)
                     </td>
                 </tr>
                 <tr>
-                    <td class="tdstat"><?php echo ($lang['ADMIN_PARAMS_GAME_SPEED']); ?><?php echo help("profile_speed_uni"); ?></td>
-                    <td class="tdvalue"><input name="speed_uni" id="speed_uni" type="text" size="5" maxlength="2" value="<?php echo $speed_uni; ?>"></td>
+                    <td class="tdstat"><?= ($lang['ADMIN_PARAMS_GAME_SPEED']) ?><?= help("profile_speed_uni") ?></td>
+                    <td class="tdvalue"><input name="speed_uni" id="speed_uni" type="text" size="5" maxlength="2" value="<?= $speed_uni ?>"></td>
                 </tr>
                 <tr>
-                    <td class="tdstat"><?php echo ($lang['ADMIN_PARAMS_GAME_SPEED_FLEET_PEACEFUL']); ?><?php echo help("profile_speed_fleet_peaceful"); ?></td>
-                    <td class="tdvalue"><input name="speed_fleet_peaceful" id="speed_fleet_peaceful" type="text" size="5" maxlength="2" value="<?php echo $speed_fleet_peaceful; ?>"></td>
+                    <td class="tdstat"><?= ($lang['ADMIN_PARAMS_GAME_SPEED_FLEET_PEACEFUL']) ?><?= help("profile_speed_fleet_peaceful") ?></td>
+                    <td class="tdvalue"><input name="speed_fleet_peaceful" id="speed_fleet_peaceful" type="text" size="5" maxlength="2" value="<?= $speed_fleet_peaceful ?>"></td>
                 </tr>
                 <tr>
-                    <td class="tdstat"><?php echo ($lang['ADMIN_PARAMS_GAME_SPEED_FLEET_WAR']); ?><?php echo help("profile_speed_fleet_war"); ?></td>
-                    <td class="tdvalue"><input name="speed_fleet_war" id="speed_fleet_war" type="text" size="5" maxlength="2" value="<?php echo $speed_fleet_war; ?>"></td>
+                    <td class="tdstat"><?= ($lang['ADMIN_PARAMS_GAME_SPEED_FLEET_WAR']) ?><?= help("profile_speed_fleet_war") ?></td>
+                    <td class="tdvalue"><input name="speed_fleet_war" id="speed_fleet_war" type="text" size="5" maxlength="2" value="<?= $speed_fleet_war ?>"></td>
                 </tr>
                 <tr>
-                    <td class="tdstat"><?php echo ($lang['ADMIN_PARAMS_GAME_SPEED_FLEET_HOLDING']); ?><?php echo help("profile_speed_fleet_holding"); ?></td>
-                    <td class="tdvalue"><input name="speed_fleet_holding" id="speed_fleet_holding" type="text" size="5" maxlength="2" value="<?php echo $speed_fleet_holding; ?>"></td>
+                    <td class="tdstat"><?= ($lang['ADMIN_PARAMS_GAME_SPEED_FLEET_HOLDING']) ?><?= help("profile_speed_fleet_holding") ?></td>
+                    <td class="tdvalue"><input name="speed_fleet_holding" id="speed_fleet_holding" type="text" size="5" maxlength="2" value="<?= $speed_fleet_holding ?>"></td>
                 </tr>
                 <tr>
-                    <td class="tdstat"><?php echo ($lang['ADMIN_PARAMS_GAME_DDR']); ?><?php echo help("profile_ddr"); ?></td>
+                    <td class="tdstat"><?= ($lang['ADMIN_PARAMS_GAME_DDR']) ?><?= help("profile_ddr") ?></td>
                     <td class="tdvalue"><input name="ddr" value="1" type="checkbox" <?php print ($ddr == 1) ? ' checked' : ''  ?>></td>
                 </tr>
                 <tr>
-                    <td class="tdstat"><?php echo ($lang['ADMIN_PARAMS_GAME_ASTRO']); ?><?php echo help("astro_strict"); ?></td>
+                    <td class="tdstat"><?= ($lang['ADMIN_PARAMS_GAME_ASTRO']) ?><?= help("astro_strict") ?></td>
                     <td class="tdvalue"><input name="astro_strict" value="1" type="checkbox" <?php print ($astro_strict == 1) ? ' checked' : ''  ?>></td>
                 </tr>
                 <tr>
-                    <td class="tdstat"><?php echo ($lang['ADMIN_PARAMS_GAME_DONUT_SYSTEM']); ?><?php echo help("donutSystem"); ?></td>
+                    <td class="tdstat"><?= ($lang['ADMIN_PARAMS_GAME_DONUT_SYSTEM']) ?><?= help("donutSystem") ?></td>
                     <td class="tdvalue"><input name="donutSystem" value="1" type="checkbox" <?php print ($donutSystem == 1) ? ' checked' : ''  ?>></td>
                 </tr>
                 <tr>
-                    <td class="tdstat"><?php echo ($lang['ADMIN_PARAMS_GAME_DONUT_GALAXY']); ?><?php echo help("donutGalaxy"); ?></td>
+                    <td class="tdstat"><?= ($lang['ADMIN_PARAMS_GAME_DONUT_GALAXY']) ?><?= help("donutGalaxy") ?></td>
                     <td class="tdvalue"><input name="donutGalaxy" value="1" type="checkbox" <?php print ($donutGalaxy == 1) ? ' checked' : ''  ?>></td>
                 </tr>
             </tbody>
         <?php endif; ?>
         <thead>
             <tr>
-                <th colspan="2"><?php echo ($lang['ADMIN_PARAMS_CACHE']); ?></th>
+                <th colspan="2"><?= ($lang['ADMIN_PARAMS_CACHE']) ?></th>
             </tr>
         </thead>
         <tbody>
             <tr>
-                <td class="tdstat"><?php echo ($lang['ADMIN_PARAMS_CACHE_RESET']); ?></td>
+                <td class="tdstat"><?= ($lang['ADMIN_PARAMS_CACHE_RESET']) ?></td>
                 <td class="tdvalue"><input name="regenere_cache" type="checkbox" value="0" /></td>
             </tr>
             <tr>
-                <td class="tdstat"><?php echo ($lang['ADMIN_PARAMS_CACHE_DURATION_CONFIG']); ?> <?php echo help("config_cache"); ?> 
+                <td class="tdstat"><?= ($lang['ADMIN_PARAMS_CACHE_DURATION_CONFIG']) ?> <?= help("config_cache") ?>
                     <br />
-                    <i class="og-warning"><?php echo ($lang['ADMIN_PARAMS_CACHE_DURATION_NOTICE']); ?></i>
+                    <i class="og-warning"><?= ($lang['ADMIN_PARAMS_CACHE_DURATION_NOTICE']) ?></i>
                 </td>
-                <td class="tdvalue"><input type="text" name="config_cache" maxlength="10" size="10" value="<?php echo $config_cache; ?>">
+                <td class="tdvalue"><input type="text" name="config_cache" maxlength="10" size="10" value="<?= $config_cache ?>">
                 </td>
             </tr>
             <tr>
-                <td class="tdstat"><?php echo ($lang['ADMIN_PARAMS_CACHE_DURATION_MOD']); ?><?php echo help("mod_cache"); ?></a></td>
-                <td class="tdvalue"><input type="text" name="mod_cache" maxlength="10" size="10" value="<?php echo $mod_cache; ?>"></td>
+                <td class="tdstat"><?= ($lang['ADMIN_PARAMS_CACHE_DURATION_MOD']) ?><?= help("mod_cache") ?></a></td>
+                <td class="tdvalue"><input type="text" name="mod_cache" maxlength="10" size="10" value="<?= $mod_cache ?>"></td>
             </tr>
 
         </tbody>
         <thead>
             <tr>
-                <th colspan="2"><?php echo ($lang['ADMIN_PARAMS_DEBUG']); ?></th>
+                <th colspan="2"><?= ($lang['ADMIN_PARAMS_DEBUG']) ?></th>
             </tr>
         </thead>
         <tbody>
             <tr>
-                <td class="tdstat"><?php echo ($lang['ADMIN_PARAMS_DEBUG_PHP']); ?><br />
+                <td class="tdstat"><?= ($lang['ADMIN_PARAMS_DEBUG_PHP']) ?><br />
 
-                    <i class="og-warning"><?php echo ($lang['ADMIN_PARAMS_DEBUG_PHP_NOTICE']); ?></i>
+                    <i class="og-warning"><?= ($lang['ADMIN_PARAMS_DEBUG_PHP_NOTICE']) ?></i>
                 </td>
-                <td class="tdvalue"><input name="log_phperror" type="checkbox" value="1" <?php echo $log_phperror; ?>></td>
+                <td class="tdvalue"><input name="log_phperror" type="checkbox" value="1" <?= $log_phperror ?>></td>
             </tr>
             <tr>
                 <td colspan="2">
-                    <input class="og-button" type="submit" value="<?php echo ($lang['ADMIN_PARAMS_VALIDATE']); ?>">
-                    <input class="og-button og-button-warning " type="reset" value="<?php echo ($lang['ADMIN_PARAMS_CANCEL']); ?>">
+                    <input class="og-button" type="submit" value="<?= ($lang['ADMIN_PARAMS_VALIDATE']) ?>">
+                    <input class="og-button og-button-warning " type="reset" value="<?= ($lang['ADMIN_PARAMS_CANCEL']) ?>">
                 </td>
             </tr>
         </tbody>
