@@ -1,4 +1,4 @@
-<?php
+<?php global $lang;
 /**
  * Page Login
  * @package OGSpy
@@ -18,7 +18,8 @@ if (!isset($goto)) {
 }
 $enable_register_view = isset($server_config['enable_register_view']) ? $server_config['enable_register_view'] : 0;
 
-require_once('views/page_header_2.php');
+require_once 'views/page_header_2.php';
+require_once dirname(__DIR__) . '/includes/token.php';
 ?>
 
 <div class="page_login">
@@ -56,12 +57,12 @@ require_once('views/page_header_2.php');
         <div class="og-login-group">
                   <p>
             <?php echo $lang['LOGIN_ACCOUNT_REQUEST_DESC']; ?>
-        </p>  
+        </p>
         </div>
 
 
-        <input class="og-button"  type="button" value="<?= $lang['LOGIN_ACCOUNT_REQUEST_BUTTON'] ?>" onclick="window.open('<?= $server_config['register_forum'] ?>');" />
+        <input class="og-button"  type="button" value="<?= $lang['LOGIN_ACCOUNT_REQUEST_BUTTON'] ?>" onclick="window.open('<?= isset($server_config['register_forum']) ? $server_config['register_forum'] : '#' ?>');" />
 
     </div>
 
-<?php require_once('views/page_tail_2.php'); ?>
+<?php require_once 'views/page_tail_2.php'; ?>
